Renamed quotes package.
