+++ src/main/scala/org/economicsl/auctions/singleunit/quoting/QuotePolicy.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions.singleunit.quotes
+package org.economicsl.auctions.singleunit.quoting
 
 import org.economicsl.auctions.Tradable
 import org.economicsl.auctions.quotes._
